requirements.txt
openai-whisper
streamlit==1.28.2
torch>=2.0.0
torchaudio>=2.0.0
numba>=0.59.0
numpy>=1.26.0
tqdm
more-itertools
transformers>=4.19.0
ffmpeg-python==0.2.0
setuptools-rust
